refactor(admin): modularize plan form, catalog requests, and business modals (Phase 4)
